requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
bson
pymongo
passlib
bcrypt
